fix(codegen): restore fragments.graphql for DatasetExplorer to break codegen/test loop
The merge resolution deleted fragments.graphql and inlined DatasetExplorer_file
as a const variable in the .tsx component. Codegen does not generate a standalone
*FragmentDoc export for a gql variable that is both assigned to fragments.file and
interpolated into another gql template, causing DatasetExplorer_FileFragmentDoc to
be missing from the generated file while still being referenced — requiring a manual
patch after every codegen run.

Restores fragments.graphql, reverts DatasetExplorer.tsx to the pre-merge pattern
(no inline const fileFragment, type imported from fragments.generated), and
regenerates all affected files.
