+++ frontend/src/datasets/features/DatasetExplorer/fragments.graphql
@@ -1,0 +1,14 @@
+fragment DatasetExplorer_file on DatasetVersionFile {
+      id
+      filename
+      createdAt
+      createdBy {
+        displayName
+      }
+      ...DownloadVersionFile_file
+      ...DatasetVersionFileSample_file
+      ...DatasetVersionFileColumns_file
+      contentType
+      size
+      uri
+    }
